Desktop/AutoOrder/ch1/requirements.txt
django_storages
boto3
dj-database-url